delete should not delete unversioned file when --keep-local is specified.

+++ svnkit/src/main/java/org/tmatesoft/svn/core/internal/wc2/ng/SvnNgRemove.java
@@ -110,7 +110,9 @@ public class SvnNgRemove extends SvnNgOperationRunner<SvnScheduleForRemoval, Svn
                     SVNErrorMessage err = SVNErrorMessage.create(SVNErrorCode.BAD_FILENAME, "''{0}'' does not exist", path);
                     SVNErrorManager.error(err, SVNLogType.WC);
                 }
-                SVNFileUtil.deleteAll(path, handler);
+                if (!keepLocal) {
+                    SVNFileUtil.deleteAll(path, handler);
+                }
                 return;
             }
             throw e;
